+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * :sparkles: Adds LifecycleBroadcastReceiver
 * :white_check_mark: Perfect Dialogx testing
 * :sparkles: Adds BundleBuilder
+* :fire: Fragmentx instance rename to instantiate, instanceOrigin rename to instantiateOrigin
 
 ## v0.6.0